+++ examples/auto-relay/dialer.js
@@ -3,6 +3,7 @@ import { webSockets } from '@libp2p/websockets'
 import { noise } from '@chainsafe/libp2p-noise'
 import { mplex } from '@libp2p/mplex'
 import { multiaddr } from '@multiformats/multiaddr'
+import { yamux } from '@chainsafe/libp2p-yamux/dist/src'
 import { circuitRelayTransport } from 'libp2p/circuit-relay'
 import { identifyService } from 'libp2p/identify'
 
@@ -21,6 +22,7 @@ async function main () {
       noise()
     ],
     streamMuxers: [
+      yamux(),
       mplex()
     ],
     services: {