+++ examples/delegated-routing/src/libp2p-bundle.js
@@ -5,13 +5,14 @@ import { createLibp2p } from 'libp2p'
 import { webSockets } from '@libp2p/websockets'
 import { webRTCStar } from '@libp2p/webrtc-star'
 import { mplex } from '@libp2p/mplex'
+import { yamux } from '@chainsafe/libp2p-yamux'
 import { noise } from '@chainsafe/libp2p-noise'
 import { delegatedPeerRouting } from '@libp2p/delegated-peer-routing'
 import { delegatedContentRouting } from '@libp2p/delegated-content-routing'
 import { create as createIpfsHttpClient } from 'ipfs-http-client'
 import { circuitRelayTransport } from 'libp2p/circuit-relay'
 
-export default function Libp2pBundle ({peerInfo, peerBook}) {
+export default function Libp2pBundle ({ peerInfo, peerBook }) {
   const wrtcstar = new webRTCStar()
   const client = createIpfsHttpClient({
     host: '0.0.0.0',
@@ -38,7 +39,7 @@ export default function Libp2pBundle ({peerInfo, peerBook}) {
       circuitRelayTransport()
     ],
     streamMuxers: [
-      mplex()
+      yamux(), mplex()
     ],
     peerDiscovery: [
       wrtcstar.discovery