+++ interop/src/index.ts
@@ -1,0 +1,3 @@
+console.log("Everything is defined in the test folder")
+
+export { }