+++ test/connection-manager/direct.spec.ts
@@ -1,38 +1,40 @@
 /* eslint-env mocha */
 
-import { expect } from 'aegir/chai'
-import sinon from 'sinon'
-import pDefer from 'p-defer'
-import delay from 'delay'
+import { yamux } from '@chainsafe/libp2p-yamux'
+import { mockConnectionGater, mockDuplex, mockMultiaddrConnection, mockUpgrader, mockConnection } from '@libp2p/interface-mocks'
+import { AbortError } from '@libp2p/interfaces/errors'
+import { EventEmitter } from '@libp2p/interfaces/events'
+import { mplex } from '@libp2p/mplex'
+import { peerIdFromString } from '@libp2p/peer-id'
+import { createFromJSON } from '@libp2p/peer-id-factory'
+import { PersistentPeerStore } from '@libp2p/peer-store'
+import { publicAddressesFirst } from '@libp2p/utils/address-sort'
 import { webSockets } from '@libp2p/websockets'
 import * as filters from '@libp2p/websockets/filters'
-import { mplex } from '@libp2p/mplex'
-import { plaintext } from '../../src/insecure/index.js'
-import type { Multiaddr } from '@multiformats/multiaddr'
 import { multiaddr } from '@multiformats/multiaddr'
-import { AbortError } from '@libp2p/interfaces/errors'
+import { expect } from 'aegir/chai'
 import { MemoryDatastore } from 'datastore-core/memory'
-import { codes as ErrorCodes } from '../../src/errors.js'
-import { publicAddressesFirst } from '@libp2p/utils/address-sort'
-import { PersistentPeerStore } from '@libp2p/peer-store'
-import { DefaultTransportManager } from '../../src/transport-manager.js'
-import { mockConnectionGater, mockDuplex, mockMultiaddrConnection, mockUpgrader, mockConnection } from '@libp2p/interface-mocks'
-import { createPeerId } from '../utils/creators/peer.js'
-import type { TransportManager } from '@libp2p/interface-transport'
-import { peerIdFromString } from '@libp2p/peer-id'
-import type { Connection } from '@libp2p/interface-connection'
-import { createLibp2p } from '../../src/index.js'
-import { DefaultConnectionManager } from '../../src/connection-manager/index.js'
-import { createFromJSON } from '@libp2p/peer-id-factory'
-import Peers from '../fixtures/peers.js'
-import { MULTIADDRS_WEBSOCKETS } from '../fixtures/browser.js'
-import type { PeerId } from '@libp2p/interface-peer-id'
+import delay from 'delay'
+import pDefer from 'p-defer'
 import { pEvent } from 'p-event'
-import { defaultComponents, Components } from '../../src/components.js'
+import sinon from 'sinon'
 import { stubInterface } from 'sinon-ts'
-import { EventEmitter } from '@libp2p/interfaces/events'
+import { defaultComponents, type Components } from '../../src/components.js'
+import { DefaultConnectionManager } from '../../src/connection-manager/index.js'
+import { codes as ErrorCodes } from '../../src/errors.js'
+import { identifyService } from '../../src/identify/index.js'
+import { createLibp2p } from '../../src/index.js'
+import { plaintext } from '../../src/insecure/index.js'
+import { DefaultTransportManager } from '../../src/transport-manager.js'
+import { MULTIADDRS_WEBSOCKETS } from '../fixtures/browser.js'
+import Peers from '../fixtures/peers.js'
+import { createPeerId } from '../utils/creators/peer.js'
+import type { DefaultIdentifyService } from '../../src/identify/identify.js'
+import type { Connection } from '@libp2p/interface-connection'
 import type { Libp2p } from '@libp2p/interface-libp2p'
-import { IdentifyService, identifyService } from '../../src/identify/index.js'
+import type { PeerId } from '@libp2p/interface-peer-id'
+import type { TransportManager } from '@libp2p/interface-transport'
+import type { Multiaddr } from '@multiformats/multiaddr'
 
 const unsupportedAddr = multiaddr('/ip4/127.0.0.1/tcp/9999')
 
@@ -229,7 +231,7 @@ describe('dialing (direct, WebSockets)', () => {
         deferredDial.reject(new AbortError())
       }
       options.signal.addEventListener('abort', onAbort)
-      return await deferredDial.promise
+      return deferredDial.promise
     })
 
     // Perform 3 multiaddr dials
@@ -341,8 +343,7 @@ describe('dialing (direct, WebSockets)', () => {
 })
 
 describe('libp2p.dialer (direct, WebSockets)', () => {
-  // const connectionGater = mockConnectionGater()
-  let libp2p: Libp2p<{ identify: IdentifyService }>
+  let libp2p: Libp2p<{ identify: unknown }>
   let peerId: PeerId
 
   beforeEach(async () => {
@@ -366,6 +367,7 @@ describe('libp2p.dialer (direct, WebSockets)', () => {
         })
       ],
       streamMuxers: [
+        yamux(),
         mplex()
       ],
       connectionEncryption: [
@@ -373,14 +375,15 @@ describe('libp2p.dialer (direct, WebSockets)', () => {
       ],
       services: {
         identify: identifyService()
-      }
+      },
+      connectionGater: mockConnectionGater()
     })
 
     if (libp2p.services.identify == null) {
       throw new Error('Identify service missing')
     }
 
-    const identifySpy = sinon.spy(libp2p.services.identify, 'identify')
+    const identifySpy = sinon.spy(libp2p.services.identify as DefaultIdentifyService, 'identify')
     const peerStorePatchSpy = sinon.spy(libp2p.peerStore, 'patch')
     const connectionPromise = pEvent(libp2p, 'connection:open')
 
@@ -409,11 +412,13 @@ describe('libp2p.dialer (direct, WebSockets)', () => {
         })
       ],
       streamMuxers: [
+        yamux(),
         mplex()
       ],
       connectionEncryption: [
         plaintext()
-      ]
+      ],
+      connectionGater: mockConnectionGater()
     })
 
     await libp2p.start()
@@ -437,11 +442,13 @@ describe('libp2p.dialer (direct, WebSockets)', () => {
         })
       ],
       streamMuxers: [
+        yamux(),
         mplex()
       ],
       connectionEncryption: [
         plaintext()
-      ]
+      ],
+      connectionGater: mockConnectionGater()
     })
 
     await libp2p.hangUp(MULTIADDRS_WEBSOCKETS[0])
@@ -456,11 +463,13 @@ describe('libp2p.dialer (direct, WebSockets)', () => {
         })
       ],
       streamMuxers: [
+        yamux(),
         mplex()
       ],
       connectionEncryption: [
         plaintext()
-      ]
+      ],
+      connectionGater: mockConnectionGater()
     })
 
     await libp2p.start()