+++ test/utils/base-options.ts
@@ -1,15 +1,16 @@
-import { tcp } from '@libp2p/tcp'
+import { yamux } from '@chainsafe/libp2p-yamux'
 import { mplex } from '@libp2p/mplex'
-import { plaintext } from '../../src/insecure/index.js'
-import type { Libp2pOptions } from '../../src'
-import mergeOptions from 'merge-options'
-import type { ServiceMap } from '@libp2p/interface-libp2p'
+import { tcp } from '@libp2p/tcp'
 import { webSockets } from '@libp2p/websockets'
 import * as filters from '@libp2p/websockets/filters'
+import mergeOptions from 'merge-options'
 import { circuitRelayTransport } from '../../src/circuit-relay/index.js'
+import { plaintext } from '../../src/insecure/index.js'
 import { MULTIADDRS_WEBSOCKETS } from '../fixtures/browser.js'
+import type { Libp2pOptions } from '../../src'
+import type { ServiceMap } from '@libp2p/interface-libp2p'
 
-export function createBaseOptions <T extends ServiceMap = {}> (...overrides: Array<Libp2pOptions<T>>): Libp2pOptions<T> {
+export function createBaseOptions <T extends ServiceMap = Record<string, unknown>> (...overrides: Array<Libp2pOptions<T>>): Libp2pOptions<T> {
   const options: Libp2pOptions = {
     addresses: {
       listen: [`${MULTIADDRS_WEBSOCKETS}/p2p-circuit`]
@@ -22,6 +23,7 @@ export function createBaseOptions <T extends ServiceMap = {}> (...overrides: Arr
       circuitRelayTransport()
     ],
     streamMuxers: [
+      yamux(),
       mplex()
     ],
     connectionEncryption: [